graphql api is now protected with jwt auth. Also, moved all the mutations in schema.py to be in mutation.py
