+++ sim_registration/customers/schema.py
@@ -2,8 +2,8 @@ import graphene
 import graphql_jwt
 from graphene_django import DjangoObjectType
 from .models import Customer, SimType, IDType
-from .middleware import AuthMiddleware
-from graphql_jwt.decorators import login_required
+from .mutation import Mutation
+from graphql_jwt.decorators import login_required, token_auth
 from graphql_jwt.shortcuts import get_token
 from django.contrib.auth import authenticate, get_user_model, login, logout
 
@@ -36,60 +36,7 @@ class UserType(DjangoObjectType):
         model = get_user_model()
         fields = ('id', 'username', 'email', 'first_name', 'last_name')
 
-class CreateCustomer(graphene.Mutation):
-    class Arguments:
-        sim_type_id = graphene.Int(required=True)
-        first_name = graphene.String(required=True)
-        last_name = graphene.String(required=True)
-        mssisdn = graphene.String(required=True)
-        profession = graphene.String(required=True)
-        id_number = graphene.String(required=True)
-        id_type_id = graphene.Int(required=True)
-        id_picture = graphene.String(required=True)
-        address = graphene.String(required=True)
-        nationality = graphene.String(required=True)
-        date_created = graphene.DateTime()
-
-    ok = graphene.Boolean()
-    customer = graphene.Field(CustomerType)
-
-    def mutate(self, info, sim_type_id, first_name, last_name, mssisdn, profession, id_number, id_type_id, id_picture, address, nationality):
-        """
-            performs the mutation using django ORM
-        """
-        sim_type = SimType.objects.get(id=sim_type_id)
-        id_type = IDType.objects.get(id=id_type_id)
-
-        print("This is the sim type", sim_type, id_type)
-        customer = Customer.objects.create(
-            sim_type=sim_type,
-            first_name=first_name,
-            last_name=last_name,
-            mssisdn=mssisdn,
-            profession=profession,
-            id_number=id_number,
-            id_type=id_type,
-            id_picture=id_picture,
-            address=address,
-            nationality=nationality
-        )
-
-        # save
-        customer.save()
-        return CreateCustomer(customer=customer)
-
-class deleteCustomer(graphene.Mutation):
-    class Arguments:
-        id = graphene.Int(required=True)
-
-    ok = graphene.Boolean()
-
-    def mutate(self, info, id):
-        customer = Customer.objects.get(id=id)
-        customer.delete()
-        return deleteCustomer(ok=True)
-
-class Query(AuthMiddleware,graphene.ObjectType):
+class Query(graphene.ObjectType):
     """
         Represent the root class that
         specifies the set of operation for quering
@@ -142,13 +89,6 @@ class Login(graphene.Mutation):
             raise Exception('Invalid username or password')
 
         token = get_token(user)
-        login(info.context, user)
         return Login(token=token, user=user)
-
-class Mutation(AuthMiddleware, graphene.ObjectType):
-    #token_auth = graphql_jwt.ObtainJSONWebToken.Field()
-    create_customer = CreateCustomer.Field()
-    delete_customer = deleteCustomer.Field()
-    login = Login.Field()
 
 schema = graphene.Schema(query=Query, mutation=Mutation)
